+++ src/components/Modal/ModalType/TodoCardModal/CommentBox.tsx
@@ -1,5 +1,7 @@
-import styles from './CommentBox.module.scss';
+import { useState } from 'react';
 import { format } from 'date-fns';
+import httpClient from '@/src/apis/httpClient';
+import styles from './CommentBox.module.scss';
 
 type Comment = {
   id: number;
@@ -20,10 +22,34 @@ interface CommentBoxProps {
   onDeleteComment?: (commentId: number) => Promise<void>;
 }
 
-//댓글 수정 기능 추가해야함
 const CommentBox = ({ data, assigneeId, onDeleteComment }: CommentBoxProps) => {
+  const [commentEdit, setCommentEdit] = useState<string>(data?.content);
+  const [isModify, setIsModify] = useState<boolean>(false);
+
+  //댓글 입력창 핸들러
+  const handleCommentEditInput = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+    setCommentEdit(e.target.value);
+  };
+
+  //날짜 포맷 함수
   const formatDate = (date: string) => {
     return format(date, 'yyyy.MM.dd HH:mm');
+  };
+
+  //댓글 수정 핸들러 (로직 수정 필요)
+  const handleCommentEdit = async (commentId: number, content: string) => {
+    try {
+      const response = await httpClient.put(`/comments/${commentId}`, {
+        content: content,
+      });
+      console.log('댓글 수정 성공:', response.data);
+      window.alert('댓글이 수정되었습니다.');
+      window.location.reload();
+      ///fetchCommentData();
+      //fetchCommentData(cursorId);
+    } catch (error) {
+      console.error('댓글 수정 실패:', error);
+    }
   };
 
   return (
@@ -38,10 +64,26 @@ const CommentBox = ({ data, assigneeId, onDeleteComment }: CommentBoxProps) => {
           <span className={styles.commentAuthorNickname}>{data?.author?.nickname}</span>
           <div className={styles.commentDate}>{formatDate(data?.createdAt)}</div>
         </div>
-        <div className={styles.commentContent}>{data?.content}</div>
+        {isModify ? (
+          <>
+            <textarea className={styles.editInput} value={commentEdit} onChange={handleCommentEditInput} />
+            <div className={styles.commentMenu}>
+              <button className={styles.commentEdit} onClick={() => handleCommentEdit(data.id, commentEdit)}>
+                수정사항 저장
+              </button>
+              <button className={styles.commentDelete} onClick={() => setIsModify(false)}>
+                수정 취소
+              </button>
+            </div>
+          </>
+        ) : (
+          <div className={styles.commentContent}>{data?.content}</div>
+        )}
         {data?.author?.id === assigneeId ? ( //담당자 id가 아니라, 로그인된 id로 비교해야함;;
           <div className={styles.commentMenu}>
-            <button className={styles.commentEdit}>수정</button>
+            <button className={styles.commentEdit} onClick={() => setIsModify(true)}>
+              수정
+            </button>
             <button className={styles.commentDelete} onClick={() => onDeleteComment && onDeleteComment(data?.id)}>
               삭제
             </button>
